github: change node version of cli to 20
