+++ .github/workflows/build.yml
@@ -244,7 +244,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v2
         with:
-          node-version: "16"
+          node-version: "20"
       - run: |
           cd clients/js && make test
           make docs
