+++ clients/js/package.json
@@ -42,7 +42,7 @@
     "@solana/web3.js": "^1.22.0",
     "@terra-money/terra.js": "^3.1.9",
     "@types/config": "^3.3.0",
-    "@wormhole-foundation/sdk": "^0.6.7",
+    "@wormhole-foundation/sdk": "^0.7.0-beta.5",
     "@xpla/xpla.js": "^0.2.1",
     "algosdk": "^2.4.0",
     "aptos": "^1.3.16",
@@ -74,6 +74,6 @@
     "jest": "^29.7.0",
     "ts-jest": "^29.1.2",
     "tsx": "^4.12.0",
-    "typescript": "^4.6"
+    "typescript": "^5.4.5"
   }
 }
