+++ clients/js/src/algorand.ts
@@ -2,17 +2,11 @@ import {
   _submitVAAAlgorand,
   signSendAndConfirmAlgorand,
 } from "@certusone/wormhole-sdk/lib/esm/algorand";
-import {
-  Account,
-  Algodv2,
-  decodeAddress,
-  encodeAddress,
-  mnemonicToSecretKey,
-} from "algosdk";
+import { Account, Algodv2, mnemonicToSecretKey } from "algosdk";
 import { NETWORKS } from "./consts";
 import { Payload, impossible } from "./vaa";
 import { transferFromAlgorand } from "@certusone/wormhole-sdk/lib/esm/token_bridge/transfer";
-import { tryNativeToHexString } from "./array";
+import { tryNativeToHexString } from "./sdk/array";
 import {
   Chain,
   chainToChainId,
@@ -20,7 +14,6 @@ import {
   Network,
   toChainId,
 } from "@wormhole-foundation/sdk-base";
-import { encoding } from "@wormhole-foundation/sdk";
 
 export async function execute_algorand(
   payload: Payload,
@@ -67,9 +60,6 @@ export async function execute_algorand(
     }
     case "NFTBridge": {
       const nftContract = contracts.nftBridge.get(network, chain);
-      // NOTE: this code can safely be removed once the algorand NFT bridge is
-      // released, but it's fine for it to stay, as the condition will just be
-      // skipped once 'contracts.nft_bridge' is defined
       if (!nftContract) {
         throw new Error("NFT bridge not supported yet for Algorand");
       }
@@ -162,7 +152,6 @@ export async function transferAlgorand(
   if (!key) {
     throw Error(`No ${network} key defined for Algorand`);
   }
-  // const contracts = CONTRACTS[network].algorand;
   const client = getClient(network, rpc);
   const wallet: Account = mnemonicToSecretKey(key);
   const CORE_ID = BigInt(parseInt(contracts.coreBridge(network, "Algorand")));
@@ -207,16 +196,4 @@ function getClient(network: Network, rpc: string) {
     ALGORAND_HOST.algodPort
   );
   return client;
-}
-
-export function uint8ArrayToNativeStringAlgorand(a: Uint8Array): string {
-  return encodeAddress(a);
-}
-
-export function hexToNativeStringAlgorand(s: string): string {
-  return uint8ArrayToNativeStringAlgorand(encoding.hex.decode(s));
-}
-
-export function nativeStringToHexAlgorand(s: string): string {
-  return encoding.hex.encode(decodeAddress(s).publicKey);
 }
