+++ clients/js/src/cmds/info/contract.ts
@@ -1,7 +1,7 @@
 import yargs from "yargs";
 import { impossible } from "../../vaa";
-import { assertChain, chains, contracts } from "@wormhole-foundation/sdk-base";
-import { getNetwork } from "../../utils";
+import { contracts } from "@wormhole-foundation/sdk-base";
+import { chainToChain, getNetwork } from "../../utils";
 
 export const command = "contract <network> <chain> <module>";
 export const desc = "Print contract address";
@@ -14,7 +14,7 @@ export const builder = (y: typeof yargs) =>
     } as const)
     .positional("chain", {
       describe: "Chain to query",
-      choices: chains,
+      type: "string",
       demandOption: true,
     } as const)
     .positional("module", {
@@ -26,8 +26,7 @@ export const handler = async (
   argv: Awaited<ReturnType<typeof builder>["argv"]>
 ) => {
   const network = getNetwork(argv.network);
-  const chain = argv.chain;
-  assertChain(chain);
+  const chain = chainToChain(argv.chain);
   const module = argv["module"];
 
   let addr: string | undefined;
