+++ clients/js/src/cmds/info/rpc.ts
@@ -1,7 +1,6 @@
 import yargs from "yargs";
 import { NETWORKS } from "../../consts";
-import { assertChain, chains } from "@wormhole-foundation/sdk-base";
-import { getNetwork } from "../../utils";
+import { chainToChain, getNetwork } from "../../utils";
 
 export const command = "rpc <network> <chain>";
 export const desc = "Print RPC address";
@@ -14,13 +13,12 @@ export const builder = (y: typeof yargs) =>
     } as const)
     .positional("chain", {
       describe: "Chain to query",
-      choices: chains,
+      type: "string",
       demandOption: true,
     } as const);
 export const handler = async (
   argv: Awaited<ReturnType<typeof builder>["argv"]>
 ) => {
-  assertChain(argv.chain);
   const network = getNetwork(argv.network);
-  console.log(NETWORKS[network][argv.chain].rpc);
+  console.log(NETWORKS[network][chainToChain(argv.chain)].rpc);
 };
